notebook: add sagemaker training example notebook
This includes examples of how to train with SageMaker locally, submit a
training job from a local notebook to AWS and run from a SageMaker notebook.
It also runs an example HPO on SageMaker and deploys the best model. A new function
was added to src/visualize to make a prediction using a deployed endpoint andplot the results.